backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-pptx==1.0.2
pydantic==2.9.1
requests==2.32.3
